Fix printf statements.

+++ teshsuite/smpi/mpich-test/pt2pt/testtypes.c
@@ -11,21 +11,21 @@ int main( int argc, char **argv )
 
 
   MPI_Type_extent(MPI_INT, &i_extent);
-  printf("MPI_Type_extent (MPI_INT) = %ld\n", i_extent);
+  printf("MPI_Type_extent (MPI_INT) = %ld\n", (long)i_extent);
 
   MPI_Type_size(MPI_INT, &i_size);
   printf("MPI_Type_size (MPI_INT) = %d\n", i_size);
 
 
   MPI_Type_extent(MPI_UNSIGNED, &i_extent);
-  printf("MPI_Type_extent (MPI_UNSIGNED) = %ld\n", i_extent);
+  printf("MPI_Type_extent (MPI_UNSIGNED) = %ld\n", (long)i_extent);
 
   MPI_Type_size(MPI_UNSIGNED, &i_size);
   printf("MPI_Type_size (MPI_UNSIGNED) = %d\n", i_size);
 
 #if defined(HAVE_LONG_DOUBLE)
   MPI_Type_extent(MPI_LONG_DOUBLE, &i_extent);
-  printf("MPI_Type_extent (MPI_LONG_DOUBLE) = %ld\n", i_extent);
+  printf("MPI_Type_extent (MPI_LONG_DOUBLE) = %ld\n", (long)i_extent);
 
   MPI_Type_size(MPI_LONG_DOUBLE, &i_size);
   printf("MPI_Type_size (MPI_LONG_DOUBLE) = %d\n", i_size);
